+++ test/integration/check_log/find_marker.py
@@ -9,7 +9,6 @@
         test/integration/check_log/find_marker.py
 
     Arguments:
-        None
 
 """
 
@@ -31,7 +30,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -43,7 +41,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Integration testing initilization.
@@ -59,7 +57,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -78,7 +75,6 @@ class UnitTest(unittest.TestCase):
         Description:  Does not find the line marker.
 
         Arguments:
-            None
 
         """
 
@@ -98,7 +94,6 @@ class UnitTest(unittest.TestCase):
         Description:  Finds the line marker.
 
         Arguments:
-            None
 
         """