+++ test/unit/check_log/fetch_log_stdin.py
@@ -9,7 +9,6 @@
         test/unit/check_log/fetch_log_stdin.py
 
     Arguments:
-        None
 
 """
 
@@ -33,7 +32,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -45,7 +43,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Unit testing initilization.
@@ -63,7 +61,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -82,7 +79,6 @@ class UnitTest(unittest.TestCase):
         Description:  Passing in no lines.
 
         Arguments:
-            mock_chk -> Mock Ref:  check_log.full_chk
 
         """
 
@@ -100,7 +96,6 @@ class UnitTest(unittest.TestCase):
         Description:  Passing in multiple lines.
 
         Arguments:
-            mock_chk -> Mock Ref:  check_log.full_chk
 
         """
 
@@ -119,8 +114,6 @@ class UnitTest(unittest.TestCase):
         Description:  Single line from standard - no find.
 
         Arguments:
-            mock_chk -> Mock Ref:  check_log.full_chk
-            mock_find -> Mock Ref:  check_log.find_marker_array
 
         """
 
@@ -139,7 +132,6 @@ class UnitTest(unittest.TestCase):
         Description:  Single line from standard - found.
 
         Arguments:
-            mock_chk -> Mock Ref:  check_log.full_chk
 
         """