+++ test/unit/check_log/fetch_marker_entry.py
@@ -9,7 +9,6 @@
         test/unit/check_log/fetch_marker_entry.py
 
     Arguments:
-        None
 
 """
 
@@ -31,7 +30,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -43,7 +41,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Unit testing initilization.
@@ -59,7 +57,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -76,7 +73,6 @@ class UnitTest(unittest.TestCase):
         Description:  Return marker with single line.
 
         Arguments:
-            None
 
         """
 
@@ -90,7 +86,6 @@ class UnitTest(unittest.TestCase):
         Description:  Return an empty marker.
 
         Arguments:
-            None
 
         """