+++ test/unit/check_log/get_filter_data.py
@@ -9,7 +9,6 @@
         test/unit/check_log/get_filter_data.py
 
     Arguments:
-        None
 
 """
 
@@ -32,7 +31,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -44,7 +42,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Unit testing initilization.
@@ -62,7 +60,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -81,7 +78,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with no -F option in args_array.
 
         Arguments:
-            None
 
         """
 
@@ -97,7 +93,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with empty expression file.
 
         Arguments:
-            None
 
         """
 
@@ -113,7 +108,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test multiple line expression file.
 
         Arguments:
-            None
 
         """
 
@@ -129,7 +123,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test getting single line expression.
 
         Arguments:
-            None
 
         """