+++ test/unit/check_log/ignore_msgs.py
@@ -9,7 +9,6 @@
         test/unit/check_log/ignore_msgs.py
 
     Arguments:
-        None
 
 """
 
@@ -31,7 +30,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -43,7 +41,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Unit testing initilization.
@@ -63,7 +61,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -90,7 +87,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with empty log array.
 
         Arguments:
-            None
 
         """
 
@@ -103,7 +99,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with all lines removed from log array.
 
         Arguments:
-            None
 
         """
 
@@ -118,7 +113,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with two ignore lines found.
 
         Arguments:
-            None
 
         """
 
@@ -133,7 +127,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with one ignore line found.
 
         Arguments:
-            None
 
         """
 
@@ -148,7 +141,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with ignore not finding any lines.
 
         Arguments:
-            None
 
         """
 
@@ -163,7 +155,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with empty ignore array.
 
         Arguments:
-            None
 
         """