+++ test/unit/check_log/open_log.py
@@ -9,7 +9,6 @@
         test/unit/check_log/open_log.py
 
     Arguments:
-        None
 
 """
 
@@ -32,7 +31,6 @@ sys.path.append(os.getcwd())
 import check_log
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -44,7 +42,7 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Unit testing initilization.
@@ -60,7 +58,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -75,8 +72,6 @@ class UnitTest(unittest.TestCase):
         Description:  Full check returns False.
 
         Arguments:
-            mock_marker -> Mock Ref:  check_log.find_marker
-            mock_chk -> Mock Ref:  check_log.full_chk
 
         """
 
@@ -86,7 +81,7 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(check_log.open_log(self.args_array), "File_Handler2")
 
     @mock.patch("check_log.full_chk")
-    @mock.patch("check_log.open")
+    @mock.patch("check_log.gen_libs.openfile")
     def test_fill_chk_true(self, mock_open, mock_chk):
 
         """Function:  test_fill_chk_true
@@ -94,8 +89,6 @@ class UnitTest(unittest.TestCase):
         Description:  Full check returns True.
 
         Arguments:
-            mock_open -> Mock Ref:  check_log.open
-            mock_chk -> Mock Ref:  check_log.full_chk
 
         """