+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.2.1] - 2019-06-18
+- Added capability to open compressed (e.g. .gz) files.
+
+### Changed
+- open_log:  Replace open call with gen_libs.openfile call.
+- find_marker:  Replace open calls with gen_libs.openfile calls.
+- get_ignore_msgs:  Replace open call with gen_libs.openfile call.
+- fetch_log:  Replace open call with gen_libs.openfile call.
+- get_filter_data:  Replace open call with gen_libs.openfile call.
+
+
 ## [2.2.0] - 2019-06-11
 - Added ability to do keyword searching.